build: bump various versions (#137)
This updates Go, golangci-lint, goreleaser, and some action definitions,
as well as whatever `go get` wants to update on its own.

To keep the diff more reasonable, I'm punting on all golangci-lint fixes
for now with the `new: true` configuration, which only checks for issues
introduced in `HEAD`. Hopefully this is short-lived.

Signed-off-by: Rob Liebowitz <rliebowitz@morningconsult.com>