ci: use default github.token for bumps
I was getting a 401 on bump, which means there might be something wrong
with the credentials we were previously trying to use. But it appears
the token that the github-push-action opts into by default should
already have permissions to do this, according to how our repo is
configured. So let's give that a go instead.

Signed-off-by: Rob Liebowitz <rliebowitz@morningconsult.com>