ci: swap back to personal access token
My new theory is that the issue was that the maintainers group (which
includes Dilan, the owner of the PAT used) was only granted "write"
access, but needed "maintain" access. This is how
morningconsult/docker-credential-vault-login was functional, because in
that repo Dilan was explicitly a maintainer.

So I have tweaked the settings and am now re-applying his token, which I
now believe is necessary based on the `github-push-action` docs.

Signed-off-by: Rob Liebowitz <rliebowitz@morningconsult.com>