refactor: reduce cognitive complexity (#142)
This removes the final set of golangci-lint issues that was barring us
from only linting against "new" issues.

I did unfortunately add one nolint directive, but the function itself is
a deep part of a concurrent process, and as far as I can tell not
covered by tests. So I figured it's probably safer to leave it alone,
especially if the options I'm considering would potentially also be
changes in behavior.

Signed-off-by: Rob Liebowitz <rliebowitz@morningconsult.com>