ci: group dependencies into monthly PRs
The current behavior is, every day at 10:00, we get a PR for every
update. For this sort of tool, that's pretty unnecessary.

This groups all Go version updates into a single PR, once per month.
The reason we can be comfortable with this cadence is that dependabot
treats version updates separately from security updates, and because
there's no urgency in non-security related updates, a monthly cadence is
a lot more on pace for our level of active maintenance.

Also clean up the properties we don't need to configure.

Signed-off-by: Rob Liebowitz <rliebowitz@morningconsult.com>